Add the orocos-bfl (Bayesian Filtering Library).
This adds a builder and interface for the orocos-bfl library.  This library
provides a number of bayesian filters.  For the linear case, it provides a
generic interface for Kalman filtering, but it also provides Extended
Kalman Filters and Particle Filters for the nonlinear case.  The library
also provides generic interaces for smoothing (e.g. Forward-Backward
filtering).